Implement game class, add GameTest(WIP)

+++ src/main/java/game/actions/Action.java
@@ -2,6 +2,6 @@ package game.actions;
 
 public interface Action {
 	public ActionType getActionType();
-	public double getMoveValue();
-	public double getRelativeMoveValue();
+	public long getMoveValue();
+	public long getRelativeMoveValue();
 }
